subsystems/gimbal: Reorganize and rearchitect `GimbalSubsystem`
This patch is a total redesign of the gimbal subsystem class. To do
this, a few key changes were made:

First of all, I removed the PID controller from the subsystem code
entirely. I thought that it would be best to put it in the gimbal
controller code rather than in the subsystem code itself, so I removed
it. This can easily be changed later if we decide that it should be in
the subsystem.

Finally, I changed the way we do angle targeting. Rather than
immediatly setting the motor output, we just update the state of the
target angle and provide an interface with which you can set the output
of the motors. This design further reinforces the idea that it's the
gimbal controller that is resposible for doing all output calculations,
rather than the subsystem code. Again, if we want to redesign this
later, that's certainly possible.