documentation +streamline
Reverted yesterday's changes to the ForAllSlideMotor functions. planning to keep the usage of the enum MotorIndex

added documentation to the slide constants for engineer

updated the refresh() function in slide.cpp to work for n motors. included an array + function to handle encoder display updates that were previously tracked with respected x/z variables